// packages/server/src/metric/enum/metric-type.enum.ts
import { registerEnumType } from '@nestjs/graphql';

export enum MetricTypeEnum {
  SHOPIFY_ORDERED_PRODUCT = 'SHOPIFY_ORDERED_PRODUCT',
  SHOPIFY_REFUNDED_ORDER = 'SHOPIFY_REFUNDED_ORDER',
  SHOPIFY_PLACED_ORDER = 'SHOPIFY_PLACED_ORDER',
  SHOPIFY_CHECKOUT_STARTED = 'SHOPIFY_CHECKOUT_STARTED',
  SHOPIFY_FULFILLED_ORDER = 'SHOPIFY_FULFILLED_ORDER',
  SHOPIFY_CANCELLED_ORDER = 'SHOPIFY_CANCELLED_ORDER',

  // meta values, measured from above value data.
  SHOPIFY_CANCELLED_ORDER_VALUE = 'SHOPIFY_CANCELLED_ORDER_VALUE',
  SHOPIFY_CHECKOUT_STARTED_VALUE = 'SHOPIFY_CHECKOUT_STARTED_VALUE',
  SHOPIFY_FULFILLED_ORDER_VALUE = 'SHOPIFY_FULFILLED_ORDER_VALUE',
  SHOPIFY_ORDERED_PRODUCT_VALUE = 'SHOPIFY_ORDERED_PRODUCT_VALUE',
  SHOPIFY_PLACED_ORDER_VALUE = 'SHOPIFY_PLACED_ORDER_VALUE',
  SHOPIFY_REFUNDED_ORDER_VALUE = 'SHOPIFY_REFUNDED_ORDER_VALUE',

  EMAIL_OPENED = 'EMAIL_OPENED',
  EMAIL_DELIVERED = 'EMAIL_DELIVERED',
  EMAIL_SENT = 'EMAIL_SENT',
  EMAIL_LINK_CLICKED = 'EMAIL_LINK_CLICKED',

  EMAIL_BOUNCED = 'EMAIL_BOUNCED',
  EMAIL_DROPPED = 'EMAIL_DROPPED',
  EMAIL_MARKED_AS_SPAM = 'EMAIL_MARKED_AS_SPAM',
  EMAIL_UNSUBSCRIBED = 'EMAIL_UNSUBSCRIBED',

  POST_PUBLISHED = 'POST_PUBLISHED',
  POST_UNPUBLISHED = 'POST_UNPUBLISHED',
  POST_DELETED = 'POST_DELETED',
  POST_VIEWED = 'POST_VIEWED',

  PRODUCT_VIEWED = 'PRODUCT_VIEWED',

  FORM_VIEWED = 'FORM_VIEWED',
  FORM_SUBMITTED = 'FORM_SUBMITTED',
}

registerEnumType(MetricTypeEnum, {
  name: 'MetricType',
  description: 'different types of metrics',
});